+++ thesis_core/src/main/java/thesis/network/messages/RequestFullStateDumpMsg.java
@@ -23,7 +23,7 @@ public class RequestFullStateDumpMsg extends InfrastructureMsg
    }
 
    @Override
-   public short getEncodedSize()
+   public long getEncodedSize()
    {
       return 0;
    }
